+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/MainView/Model/DailyBoxOfficeModel.swift
@@ -1,0 +1,15 @@
+//
+//  DailyBoxOfficeModel.swift
+//  34th-Sopt-Assignment
+//
+//  Created by 정민지 on 6/7/24.
+//
+
+import UIKit
+
+struct DailyBoxOfficeModel {
+    let date: Int
+    let itemsPerPage: Int
+    let multiMovieYn: String
+    let repNationCd: String
+}
