+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/MainView/Model/MovieModel.swift
