+++ 34th-Sopt-Assignment/34th-Sopt-Assignment/MainView/ViewModel/MainViewModel.swift
@@ -6,31 +6,40 @@
 //
 
 import RxSwift
+import RxCocoa
 import Foundation
 
-struct BoxOfficeResult {
-    let date: Int
-    let itemsPerPage: Int
-    let multiMovieYn: String
-    let repNationCd: String
-}
-
-class LiveViewModel {
+final class MainViewModel: ViewModel {
     private let disposeBag = DisposeBag()
     private let boxofficeManager = BoxofficeManager.shared
     
     // MARK: - Input
     struct Input {
-        let getDailyBoxOfficeTrigger: Observable<BoxOfficeResult>
+        let fetchTrigger: Observable<Void>
+        let getDailyBoxOfficeTrigger: Observable<DailyBoxOfficeModel>
     }
-    
+
     // MARK: - Output
     struct Output {
-        let dailyBoxOffice: Observable<[DailyBoxOfficeList]>
+        let posterData: Driver<[PosterModel]>
+        let movieData: Driver<[MovieModel]>
+        let liveData: Driver<[DailyBoxOfficeList]>
     }
     
-    // MARK: - Transformation
+    // MARK: - Transform
     func transform(_ input: Input) -> Output {
+        let posterData = input.fetchTrigger
+            .flatMapLatest { _ in
+                return Observable.just(PosterModel.dummy())
+            }
+            .asDriver(onErrorJustReturn: [])
+
+        let movieData = input.fetchTrigger
+            .flatMapLatest { _ in
+                return Observable.just(MovieModel.dummy())
+            }
+            .asDriver(onErrorJustReturn: [])
+
         let dailyBoxOffice = input.getDailyBoxOfficeTrigger
             .flatMapLatest { [weak self] request -> Observable<[DailyBoxOfficeList]> in
                 guard let self = self else { return Observable.just([]) }
@@ -40,12 +49,17 @@ class LiveViewModel {
                     multiMovieYn: request.multiMovieYn,
                     repNationCd: request.repNationCd
                 ).catchAndReturn([])
-            }.observe(on: MainScheduler.instance)
-        
-        return Output(dailyBoxOffice: dailyBoxOffice)
+            }
+            .observe(on: MainScheduler.instance)
+            .asDriver(onErrorJustReturn: [])
+
+        return Output(
+            posterData: posterData,
+            movieData: movieData,
+            liveData: dailyBoxOffice
+        )
     }
-    
-    // MARK: - API
+
     private func getDailyBoxOffice(targetDt: Int, itemPerPage: Int, multiMovieYn: String, repNationCd: String) -> Observable<[DailyBoxOfficeList]> {
         return Observable.create { observer in
             self.boxofficeManager.getDailyBoxOffice(key: "\(PrivacyInfoManager.moviekey)", targetDt: targetDt, itemPerPage: itemPerPage, multiMovieYn: multiMovieYn, repNationCd: repNationCd) { result in
